Remove anchor from "this link"
Given the feedback we got, we should make this tool as simple as humanly
possible. If "this link" is an actual link, we risk that people click on
it and get confused.

This was originally added to improve compatibility with Lynx. There is
an easy fix, though: Hit "^X e" when focusing the <input> field and Lynx
will throw you in you $EDITOR, where you can do whatever you want with
the field content.