Pass keys to clients
Instead of storing one global key on the server, we now create a new key
for each secret and pass it to the client. It's part of the secret's
URL.

Still, decryption happens on the server. It's better than before, but an
attacker still has the opportunity to do MITM.

This is a compromise. By doing server-side decryption, we avoid the need
of having to write portable / highly compatible JavaScript which would
decrypt client-side. Plus, extremely restricted browsers might not even
allow JS.

(The real risk is probably the user making a mistake in handling the
decrypted secret *after retrieval*. They might store it in an unsafe
place or pass it verbatim to other people, for example.)