Split the file upload form
First of all, when all forms are on one page, it feels cluttered and
overwhelming.

Second, when there is a combined form, you have to decide if you want to
use the textbox's content or the uploaded file. It is not entirely clear
to the user what will happen (the description explains it, but yeah ...)
and they cannot double check by revealing the secret. By splitting the
file upload form, we can eliminate this ambiguity.

Another option would be to split the combined form into two *tabs*
using JS. I explicitly do not want to go down the JS route, though, not
even for "simple" visual hints. Same goes for fancy CSS stuff which
would probably create clutter in screen readers or lynx.

By linking the two forms back and forth, it's basically "two tabs".